langgraph/requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=0.19.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
langgraph>=0.0.26

# Vector database
chromadb>=0.4.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
docx2txt>=0.8

# Enhanced document parsing
markitdown>=0.0.1a2

# Azure OpenAI
openai>=1.0.0

# Search and web APIs
tavily-python>=0.3.0

# Utilities
typing-extensions>=4.0.0
pydantic>=2.0.0
